only bind mount imports dir if it exists
stacker chroot doesn't import anything, and so in some cases there won't be
this bind mount dir. Let's only do the mount when it exists.

+++ run.go
@@ -15,11 +15,13 @@ func Run(sc StackerConfig, name string, command string, l *Layer, onFailure stri
 	}
 
 	importsDir := path.Join(sc.StackerDir, "imports", name)
-	err = c.bindMount(importsDir, "/stacker")
-	if err != nil {
-		return err
+	if _, err := os.Stat(importsDir); err == nil {
+		err = c.bindMount(importsDir, "/stacker")
+		if err != nil {
+			return err
+		}
+		defer os.Remove(path.Join(sc.RootFSDir, ".working", "rootfs", "stacker"))
 	}
-	defer os.Remove(path.Join(sc.RootFSDir, ".working", "rootfs", "stacker"))
 
 	err = c.bindMount("/etc/resolv.conf", "/etc/resolv.conf")
 	if err != nil {
